requirements.txt
autopep8==1.6.0
lxml==4.8.0
Pillow==9.1.0
pycodestyle==2.8.0
python-pptx==0.6.21
toml==0.10.2
XlsxWriter==3.0.3
